g Ene Ral  STor E
changed offsets to type double i hope this is not too scuffed

+++ src/CollectableComponent.java
@@ -14,7 +14,7 @@ public class CollectableComponent extends IntrinsicTileComponent
   public CollectableComponent(String name,
                               String imagePath,
                               int numProducts,
-                              int[] offsets) throws IOException {
+                              double[] offsets) throws IOException {
     super(name, imagePath, offsets);
 
     this.products = new HoldableDrop[numProducts];
