+++ src/Furniture.java
@@ -8,7 +8,7 @@
 //TODO: if we don't use this, destroy this file.
 import java.io.IOException;
 public class Furniture extends IntrinsicTileComponent {
-  public Furniture(String name, String imagePath, int[] offsets) throws IOException {
+  public Furniture(String name, String imagePath, double[] offsets) throws IOException {
     super(name, imagePath, offsets);
   }
 }
